added NMS to neural net (which required changing the code to more closely follow the example from OpenCV docs)

+++ CDNeuralNet.hpp
@@ -8,7 +8,7 @@
 struct _bbox {
     int classId;
     float confidence;
-    float x, y, w, h;
+    int x, y, w, h;
 };
